// swiyu-ecosystem-portal-ui/src/app/api/generated/model/paged-model-trust-onboarding-submission-document-list-item.ts
/**
 * SWIYU Ecosystem Portal SCS API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {PageMetadata} from './page-metadata';
import {TrustOnboardingSubmissionDocumentListItem} from './trust-onboarding-submission-document-list-item';

export interface PagedModelTrustOnboardingSubmissionDocumentListItem {
  content?: Array<TrustOnboardingSubmissionDocumentListItem>;
  page?: PageMetadata;
}